Feat: Brain판매, Brain 업그레이드 구현 / 네트워크 초기화 임시 구현
Former-commit-id: a9fa3b741ef5b4d61057ba1362da215229040934 [formerly 1b28732c3e2970b96be53efa802c0464528a21a2]
Former-commit-id: 9064bc4c5ccff764045b8836b1b7d0cf7d46693f

+++ Assets/Scripts/SingletonManager/Network/Response/DeleteSingleNetworkBrainResponse.cs
@@ -8,7 +8,7 @@ public class DeleteSingleNetworkBrainResponse
 	public int statusCode;
 	public List<AnsEquations> ansEquations;
 	public List<Distances> distances;
-	public List<int> deletedBrains;
+	public List<long> deletedBrains;
 	public TopCoeffs NP;
 	public long calcTime;
 }
